doc: Fix some "Cannot tie this..." errors
These are errors where clangqdoc finds a qdoc comment, but
it can't find anything in the API to associate it with. There
are different causes. In many cases, soneone has documented a
function that isn't in the public API so it should not be
documented in the first place. In other cases, a function is
platform-specific and we need to add || defined(Q_CLANG_QDOC)
somewhere to make it visible.

Change-Id: I2b707548109f626e20d28f084072ef78f2ee6ac5
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>